+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.IntuneDeviceCompliancePolicyAndroidWorkProfile.Tests.ps1
@@ -82,6 +82,58 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     SecurityRequireGooglePlayServices                  = $True
                     SecurityRequireUpToDateSecurityProviders           = $True
                     SecurityRequireCompanyPortalAppIntegrity           = $True
+                    MinAndroidSecurityPatchLevel                       = "2024-01-01";
+                    requiredPasswordComplexity                         = "medium";
+                    scheduledActionsForRule                            = [CimInstance[]]@(
+                            (New-CimInstance `
+                            -ClassName MSFT_scheduledActionConfigurations `
+                            -Property @{
+                                actionType = 'block'
+                                notificationTemplateId = '00000000-0000-0000-0000-000000000000'
+                                notificationMessageCcList = @('')
+                                gracePeriodHours = 0
+                            } -ClientOnly)
+                            (New-CimInstance `
+                            -ClassName MSFT_scheduledActionConfigurations `
+                            -Property @{
+                                actionType = 'notification'
+                                notificationTemplateId = 'fakestring0'
+                                notificationMessageCcList = @('00000000-0000-0000-0000-000000000000')
+                                gracePeriodHours = 24
+                            } -ClientOnly)
+                            (New-CimInstance `
+                            -ClassName MSFT_scheduledActionConfigurations `
+                            -Property @{
+                                actionType = 'pushNotification'
+                                notificationTemplateId = '00000000-0000-0000-0000-000000000000'
+                                notificationMessageCcList = @('')
+                                gracePeriodHours = 48
+                            } -ClientOnly)
+                            (New-CimInstance `
+                            -ClassName MSFT_scheduledActionConfigurations `
+                            -Property @{
+                                actionType = 'retire'
+                                notificationTemplateId = '00000000-0000-0000-0000-000000000000'
+                                notificationMessageCcList = @('')
+                                gracePeriodHours = 72
+                            } -ClientOnly)
+                            (New-CimInstance `
+                            -ClassName MSFT_scheduledActionConfigurations `
+                            -Property @{
+                                actionType = 'remoteLock'
+                                notificationTemplateId = '00000000-0000-0000-0000-000000000000'
+                                notificationMessageCcList = @('')
+                                gracePeriodHours = 96
+                            } -ClientOnly)
+                    )
+                    SecurityRequiredAndroidSafetyNetEvaluationType     = "hardwareBacked"
+                    workProfileInactiveBeforeScreenLockInMinutes       = 480
+                    workProfilePasswordExpirationInDays                = 30
+                    workProfilePasswordMinimumLength                   = 12
+                    workProfilePasswordRequiredType                    = "atLeastNumeric"
+                    workProfilePreviousPasswordBlockCount              = 5
+                    workProfileRequiredPasswordComplexity              = "high"
+                    WorkProfileRequirePassword                         = $True
                     Ensure                                             = 'Present'
                     Credential                                         = $Credential
                 }
@@ -132,6 +184,15 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     SecurityRequireGooglePlayServices                  = $True
                     SecurityRequireUpToDateSecurityProviders           = $True
                     SecurityRequireCompanyPortalAppIntegrity           = $True
+                    MinAndroidSecurityPatchLevel                       = "2024-01-01";
+                    SecurityRequiredAndroidSafetyNetEvaluationType     = "hardwareBacked"
+                    workProfileInactiveBeforeScreenLockInMinutes       = 480
+                    workProfilePasswordExpirationInDays                = 30
+                    workProfilePasswordMinimumLength                   = 12
+                    workProfilePasswordRequiredType                    = "atLeastNumeric"
+                    workProfilePreviousPasswordBlockCount              = 5
+                    workProfileRequiredPasswordComplexity              = "high"
+                    WorkProfileRequirePassword                         = $True
                     Ensure                                             = 'Present'
                     Credential                                         = $Credential
                 }
@@ -167,6 +228,15 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                             SecurityRequireUpToDateSecurityProviders           = $True
                             SecurityRequireCompanyPortalAppIntegrity           = $True
                             RoleScopeTagIds                                    = '0'
+                            MinAndroidSecurityPatchLevel                       = "2024-01-01";
+                            SecurityRequiredAndroidSafetyNetEvaluationType     = "hardwareBacked"
+                            workProfileInactiveBeforeScreenLockInMinutes       = 480
+                            workProfilePasswordExpirationInDays                = 30
+                            workProfilePasswordMinimumLength                   = 12
+                            workProfilePasswordRequiredType                    = "atLeastNumeric"
+                            workProfilePreviousPasswordBlockCount              = 5
+                            workProfileRequiredPasswordComplexity              = "high"
+                            WorkProfileRequirePassword                         = $True
                         }
                     }
                 }
@@ -212,7 +282,16 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     SecurityRequireSafetyNetAttestationCertifiedDevice = $True
                     SecurityRequireGooglePlayServices                  = $True
                     SecurityRequireUpToDateSecurityProviders           = $True
-                    SecurityRequireCompanyPortalAppIntegrity           = $True
+                    SecurityRequireCompanyPortalAppIntegrity           = $True                   
+                    MinAndroidSecurityPatchLevel                       = "2024-01-01";
+                    SecurityRequiredAndroidSafetyNetEvaluationType     = "hardwareBacked"
+                    workProfileInactiveBeforeScreenLockInMinutes       = 480
+                    workProfilePasswordExpirationInDays                = 30
+                    workProfilePasswordMinimumLength                   = 12
+                    workProfilePasswordRequiredType                    = "atLeastNumeric"
+                    workProfilePreviousPasswordBlockCount              = 5
+                    workProfileRequiredPasswordComplexity              = "high"
+                    WorkProfileRequirePassword                         = $True
                     Ensure                                             = 'Present'
                     Credential                                         = $Credential
                 }
@@ -248,6 +327,16 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                             SecurityRequireUpToDateSecurityProviders           = $True
                             SecurityRequireCompanyPortalAppIntegrity           = $True
                             RoleScopeTagIds                                    = '0'
+
+                            MinAndroidSecurityPatchLevel                       = "2024-01-01";
+                            SecurityRequiredAndroidSafetyNetEvaluationType     = "hardwareBacked"
+                            workProfileInactiveBeforeScreenLockInMinutes       = 480
+                            workProfilePasswordExpirationInDays                = 30
+                            workProfilePasswordMinimumLength                   = 12
+                            workProfilePasswordRequiredType                    = "atLeastNumeric"
+                            workProfilePreviousPasswordBlockCount              = 5
+                            workProfileRequiredPasswordComplexity              = "high"
+                            WorkProfileRequirePassword                         = $True
                         }
                     }
                 }